Moved removeBlock and restoreBlock funtions to the BlockHelper.
