--- packages/schema-blocks/src/blocks/warning-block/removeBlock.ts
@@ -1,10 +1,0 @@
-import { dispatch } from "@wordpress/data";
-
-/**
- * Removes the block with the given ID.
- *
- * @param {string} clientId The ID of the block to remove.
- */
-export default function removeBlock( clientId: string ) {
-	dispatch( "core/editor" ).removeBlock( clientId );
-}
